Disable interpolation of config values
We don't want any (basic) interpolation of config values when writing the
BlueprintSettings.ini and when reading the CraftSettings.ini.template to
create the CraftSettings.ini. (Extended) interpolation will be done by
craft when reading the config files.

This enables using "%" in config values without causing "ValueError:
invalid interpolation syntax in '...' as position ..." when writing
the CraftSettings.ini.